Create new HeroContainer component and styles
A new HeroContainer component has been created, reworking how hero entities are rendered on the page. A new SCSS file was also added to match this new structure. The old HeroContainer component is commented out in the export index, and replaced with our newly created one. A new SVG file was also included in the assets directory.