Add features and adjust styles in HeroContainer
Several changes were made to the HeroContainer .New styles were added to the SCSS file to improve image handling and layout, including disabling user-select and pointer-events for images, adding a border to the heroImgWrapper, and adjusting the margin top for heroDescription. Moreover, a 'LeftArrow' element was added to both styling and TSX for better navigation experience, which is also responsive for smaller screens.