Refactor ImageDistance and HeroContainer logic
This commit includes improvements in the handleResize method to enable a smoother transition and relocation of images upon resizing the screen. Additionally, modifications have been made to the HeroContainer styles as well as removing redundant comments and refining the calculation of maxHeight based on mobile or desktop view. Furthermore, logging functionality has been added to assist in tracking distance values within the ImageDistance script.