Extract NavItem component to its own file
This commit migrates the NavItem component to its own file for modularity. Previously, the NavItem component's logic was located inside the NavbarDesktopV2 file. By extracting it into its own file, the code becomes easier to maintain and improve readability. This change does not impact the functionality of the NavBar component.