Update navbar styles and add overlay toggle functionality
The NavbarDesktopV2 component has been modified to include a piece of state tracking whether it's in an 'overlaid' state or not. A button has been added to the navigation bar items to toggle this state. Several styling changes were added to NavbarDesktopV2.module.scss to support this feature. Also, the element properties in the NavbarMain file have been formatted multiline for readability.