Refactor MainPage component and improve translation handling
The MainPage component has been refactored from an async function to a standard function, removing the necessity to use the useServerTranslation hook within this component. The translation functionality is now handled at a higher level in the page component which calls MainPage. The MainPage's properties type (Props) has also been modified and exported to sync with these alterations.