Update animation for GetToKnowComics module
The opacity, transform, transition, and animation of the GetToKnowComics module were moved inside a class .inView. The main purpose of this move is for better control and more efficient handling of the animation when the component enters the viewport. The adjustment has been made without affecting the module's responsiveness on mobile view.