Update Hero Card and Section layout and add translation support
The updates include modification to the layout and styles of Hero Card and Section Classified Heroes Blocks. For easier navigation and accessibility, support for translating the titles of the Hero Cards has been added. This involves fetching the titles using a translation function and passing down the translated title as