Refactor SectionPlayWithUs component and associated styles
The SectionPlayWithUs component in the frontend migration has been dramatically restructured. Code related to the rendering of the webGl and the googlePlay button in the component has been simplified and refactored. This includes changes to both the TypeScript and SCSS files. Old, unnecessary code has been removed, and new, cleaner code has replaced it to improve readability and maintainability.