Add ClassifiedHeroesBlocksProps to MainPage component
The commit includes adding `ClassifiedHeroesBlocksProps` to MainPage component and usage of the respective `title` prop in `SectionClassifiedHeroesBlocks` component. Importing `ClassifiedHeroesBlocksProps` on MainPage.tsx and passing it to `SectionClassifiedHeroesBlocks` leads to displaying title of classified heroes blocks on the main page. Styling updates and locale translations also included for the new title.