Add 'See More' button to classified heroes section
A 'See More' button has been added to the classified heroes section in the frontend. This includes style updates, translation strings for English and Finnish languages, and modifications to the heroes section layout to accommodate the newly added button. The styling details of this button have also been defined. The button is linked to the heroes route.